+++ packages/block-library/src/site-title/edit/index.js
@@ -1,24 +1,40 @@
 /**
+ * External dependencies
+ */
+import classnames from 'classnames';
+
+/**
  * WordPress dependencies
  */
 import { useEntityProp } from '@wordpress/core-data';
-import { BlockControls, RichText } from '@wordpress/block-editor';
 import { __ } from '@wordpress/i18n';
+import {
+	RichText,
+	AlignmentToolbar,
+	BlockControls,
+	__experimentalBlock as Block,
+} from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
 import LevelToolbar from './level-toolbar';
 
-export default function SiteTitleEdit( {
-	attributes: { level },
-	setAttributes,
-} ) {
+export default function SiteTitleEdit( { attributes, setAttributes } ) {
+	const { level, textAlign } = attributes;
 	const [ title, setTitle ] = useEntityProp( 'root', 'site', 'title' );
-	const tagName = level === 0 ? 'p' : `h${ level }`;
+	const tagName = 0 === level ? 'p' : `h${ level }`;
+
 	return (
 		<>
 			<BlockControls>
+				<AlignmentToolbar
+					value={ textAlign }
+					onChange={ ( nextAlign ) => {
+						setAttributes( { textAlign: nextAlign } );
+					} }
+				/>
+
 				<LevelToolbar
 					level={ level }
 					onChange={ ( newLevel ) =>
@@ -26,12 +42,17 @@ export default function SiteTitleEdit( {
 					}
 				/>
 			</BlockControls>
+
 			<RichText
-				tagName={ tagName }
+				tagName={ Block[ tagName ] }
 				placeholder={ __( 'Site Title' ) }
 				value={ title }
 				onChange={ setTitle }
+				className={ classnames( {
+					[ `has-text-align-${ textAlign }` ]: textAlign,
+				} ) }
 				allowedFormats={ [] }
+				disableLineBreaks
 			/>
 		</>
 	);
